# docs/openapi.json
{
  "openapi" : "3.1.0",
  "info" : {
    "description" : "Reserva de turnos de ascensores para reducir la congestión en el campus",
    "title" : "SmartElevate API",
    "version" : "v0"
  },
  "servers" : [ {
    "url" : "/",
    "description" : "Mismo origen"
  } ],
  "tags" : [ {
    "description" : "Verificación de conectividad",
    "name" : "ping"
  }, {
    "description" : "Edificios, núcleos y ascensores (solo lectura)",
    "name" : "elevator"
  } ],
  "paths" : {
    "/api/buildings" : {
      "get" : {
        "operationId" : "buildings",
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/BuildingResponse"
                  }
                }
              }
            },
            "description" : "OK"
          },
          "400" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Bad Request"
          },
          "500" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Internal Server Error"
          }
        },
        "summary" : "Lista los edificios con su rango de pisos y sus núcleos",
        "tags" : [ "elevator" ]
      }
    },
    "/api/elevators" : {
      "get" : {
        "operationId" : "elevators",
        "parameters" : [ {
          "description" : "Código del edificio",
          "example" : "INDEPENDENCIA",
          "in" : "query",
          "name" : "building",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "description" : "Solo los ascensores que paran en este piso (0 = PB)",
          "example" : 11,
          "in" : "query",
          "name" : "floor",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/ElevatorResponse"
                  }
                }
              }
            },
            "description" : "OK"
          },
          "400" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Bad Request"
          },
          "500" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Internal Server Error"
          }
        },
        "summary" : "Lista los ascensores activos con los pisos donde paran",
        "tags" : [ "elevator" ]
      }
    },
    "/api/ping" : {
      "get" : {
        "operationId" : "ping",
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PingResponse"
                }
              }
            },
            "description" : "OK"
          },
          "400" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Bad Request"
          },
          "500" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Internal Server Error"
          }
        },
        "summary" : "Verifica que la API responde",
        "tags" : [ "ping" ]
      }
    }
  },
  "components" : {
    "schemas" : {
      "ApiError" : {
        "type" : "object",
        "description" : "Error estándar de la API",
        "properties" : {
          "error" : {
            "type" : "string",
            "example" : "Not Found"
          },
          "message" : {
            "type" : "string",
            "example" : "Recurso no encontrado"
          },
          "path" : {
            "type" : "string",
            "example" : "/api/turns/42"
          },
          "status" : {
            "type" : "integer",
            "format" : "int32",
            "example" : 404
          },
          "timestamp" : {
            "type" : "string",
            "format" : "date-time"
          },
          "violations" : {
            "type" : "array",
            "description" : "Solo en errores de validación",
            "items" : {
              "$ref" : "#/components/schemas/FieldViolation"
            }
          }
        },
        "required" : [ "error", "message", "path", "status", "timestamp" ]
      },
      "BuildingResponse" : {
        "type" : "object",
        "description" : "Edificio con su rango de pisos y sus núcleos",
        "properties" : {
          "code" : {
            "type" : "string",
            "example" : "LIMA"
          },
          "cores" : {
            "type" : "array",
            "description" : "Núcleos del edificio, por código",
            "items" : {
              "$ref" : "#/components/schemas/CoreResponse"
            }
          },
          "maxFloor" : {
            "type" : "integer",
            "format" : "int32",
            "example" : 10
          },
          "minFloor" : {
            "type" : "integer",
            "format" : "int32",
            "example" : -4
          },
          "name" : {
            "type" : "string",
            "example" : "Lima"
          }
        },
        "required" : [ "code", "cores", "maxFloor", "minFloor", "name" ]
      },
      "CoreResponse" : {
        "type" : "object",
        "properties" : {
          "code" : {
            "type" : "string",
            "example" : "L3"
          }
        },
        "required" : [ "code" ]
      },
      "ElevatorResponse" : {
        "type" : "object",
        "description" : "Ascensor con los pisos donde para",
        "properties" : {
          "active" : {
            "type" : "boolean"
          },
          "building" : {
            "type" : "string",
            "example" : "INDEPENDENCIA"
          },
          "code" : {
            "type" : "string",
            "example" : "33"
          },
          "core" : {
            "type" : "string",
            "example" : "IND2"
          },
          "floors" : {
            "type" : "array",
            "description" : "Pisos servidos, de menor a mayor (0 = PB)",
            "example" : [ -4, -3, -2, -1, 0, 2, 3 ],
            "items" : {
              "type" : "integer",
              "format" : "int32"
            }
          },
          "usage" : {
            "type" : "string",
            "description" : "COMMON, o dedicado a docentes (TEACHERS) o a movilidad reducida (REDUCED_MOBILITY)",
            "enum" : [ "COMMON", "TEACHERS", "REDUCED_MOBILITY" ]
          }
        },
        "required" : [ "active", "building", "code", "core", "floors", "usage" ]
      },
      "FieldViolation" : {
        "type" : "object",
        "properties" : {
          "field" : {
            "type" : "string"
          },
          "message" : {
            "type" : "string"
          }
        },
        "required" : [ "field", "message" ]
      },
      "PingResponse" : {
        "type" : "object",
        "properties" : {
          "status" : {
            "type" : "string",
            "example" : "ok"
          }
        },
        "required" : [ "status" ]
      }
    }
  }
}